+++ httpx/_utils.py
@@ -6,18 +6,8 @@ import ipaddress
 import mimetypes
 import os
 import re
-import time
 import typing
-from pathlib import Path
 from urllib.request import getproxies
-
-try:
-    from sniffio import current_async_library
-except ImportError:  # pragma: nocover
-
-    def current_async_library() -> str:
-        return "asyncio"
-
 
 from ._types import PrimitiveData
 
@@ -32,33 +22,6 @@ _HTML5_FORM_ENCODING_REPLACEMENTS.update(
 _HTML5_FORM_ENCODING_RE = re.compile(
     r"|".join([re.escape(c) for c in _HTML5_FORM_ENCODING_REPLACEMENTS.keys()])
 )
-
-
-def normalize_header_key(
-    value: str | bytes,
-    lower: bool,
-    encoding: str | None = None,
-) -> bytes:
-    """
-    Coerce str/bytes into a strictly byte-wise HTTP header key.
-    """
-    if isinstance(value, bytes):
-        bytes_value = value
-    else:
-        bytes_value = value.encode(encoding or "ascii")
-
-    return bytes_value.lower() if lower else bytes_value
-
-
-def normalize_header_value(value: str | bytes, encoding: str | None = None) -> bytes:
-    """
-    Coerce str/bytes into a strictly byte-wise HTTP header value.
-    """
-    if isinstance(value, bytes):
-        return value
-    if not isinstance(value, str):
-        raise TypeError(f"Header value must be str or bytes, not {type(value)}")
-    return value.encode(encoding or "ascii")
 
 
 def primitive_value_to_str(value: PrimitiveData) -> str:
@@ -97,18 +60,6 @@ def format_form_param(name: str, value: str) -> bytes:
 
     value = _HTML5_FORM_ENCODING_RE.sub(replacer, value)
     return f'{name}="{value}"'.encode()
-
-
-def get_ca_bundle_from_env() -> str | None:
-    if "SSL_CERT_FILE" in os.environ:
-        ssl_file = Path(os.environ["SSL_CERT_FILE"])
-        if ssl_file.is_file():
-            return str(ssl_file)
-    if "SSL_CERT_DIR" in os.environ:
-        ssl_path = Path(os.environ["SSL_CERT_DIR"])
-        if ssl_path.is_dir():
-            return str(ssl_path)
-    return None
 
 
 def parse_header_links(value: str) -> list[dict[str, str]]:
@@ -294,33 +245,6 @@ def peek_filelike_length(stream: typing.Any) -> int | None:
             return None
 
     return length
-
-
-class Timer:
-    async def _get_time(self) -> float:
-        library = current_async_library()
-        if library == "trio":
-            import trio
-
-            return trio.current_time()
-        else:
-            import asyncio
-
-            return asyncio.get_event_loop().time()
-
-    def sync_start(self) -> None:
-        self.started = time.perf_counter()
-
-    async def async_start(self) -> None:
-        self.started = await self._get_time()
-
-    def sync_elapsed(self) -> float:
-        now = time.perf_counter()
-        return now - self.started
-
-    async def async_elapsed(self) -> float:
-        now = await self._get_time()
-        return now - self.started
 
 
 class URLPattern: